WIP: Initial implementation of `sam build` command
