+++ samcli/cli/command.py
@@ -16,6 +16,7 @@ _SAM_CLI_COMMAND_PACKAGES = {
     "samcli.commands.deploy",
     "samcli.commands.package",
     "samcli.commands.logs",
+    "samcli.commands.build"
 }
